Split build.js into buildFuncs.js so the functions can be run in other files.
